requirements.txt
pyswip==0.3.3
Pillow==10.0.1
